changed controller to cleaneer

+++ interfaces-persistence/src/main/java/ar/edu/itba/paw/interfaces/persistence/EventResponseDao.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface EventResponseDao {
     EventResponse create(long userId, String username, long eventId, String message, LocalDateTime dateTime);
     List<EventResponse> listAllFromEvent(long eventId);
-    long getCount(long eventId);
+    int getCount(long eventId);
     Page<EventResponse> listAllFromEvent(long eventId, int page, int size);
 //    List<User> listAllUsersResponders(long eventId); //@todo no se usa el final, podria borrar
 //    List<User> listAllUsersRespondersMinusUsers(long eventId, List<Long> user_ids);
